VC8 warning fix
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@21486 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/tools/lighter/raytracer.cpp
@@ -219,7 +219,7 @@ bool csRaytracer::IntersectTriangles (const csKDTreeNode* node, const csRay &ray
                                       csHitPoint &hit, bool earlyExit)
 {
   //assume valid input
-  uint nIdx, nMax;
+  size_t nIdx, nMax;
   nMax = node->numberOfPrimitives;
   bool haveHit = false;
   csMeshPatchAccStruct **primList = (csMeshPatchAccStruct**)csKDTreeNodeH::GetPointer (node);
